Move tests out of deserializer
